Remove unused code and reorganize components directory
This update eliminates several unused files such as `auth.ts`, `Header.jsx`, and other deprecated components. Additionally, it reorganizes the components directory, moving relevant files to appropriate folders like `client` and `dashboard`, for improved structure and maintainability.
