+++ app/layout.jsx
@@ -45,11 +45,12 @@ function ThemeWrapper({ children }) {
   }, []);
 
   return (
-    <html lang="en" data-theme={selectedTheme} suppressHydrationWarning className={`font-family-sans tracking-wider transition-all duration-500`}>
+    <html lang="en" data-theme={selectedTheme} suppressHydrationWarning className={`font-family-sans`}>
     <head>
-      <title></title>
+      <title>EDNAV Success Tracker</title>
       <meta charSet="utf-8" />
       <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="icon" href="/favicon.ico" />
     </head>
     <body className={`font-serif`}>{isMounted && children}</body>
     </html>
