+++ components/layout/SearchField.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useFepsLeft} from '../contexts/FepsLeftContext';
+import {useFepsLeft} from '../../contexts/FepsLeftContext';
 import { X } from 'lucide-react';
 
 function SearchField() {
